Added support for webhooks.
Now external services can be notified when models in the CCXcon
happen. Requests are signed with a shared secret set on a per-client
basis, which can be used to validate the incoming request actually came
from the CCXCon.

Implementors note: When using this comparison, refer to the test for the
proper way to do that. Importantly, we shouldn't use == to compare the
hashes, but a constant time comparison.